refactor: abstract UI logic into TEA
